LayoutTests/imported/w3c/web-platform-tests/html/semantics/invokers/idlharness.tentative-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Element includes ParentNode: member names are unique
PASS Element includes NonDocumentTypeChildNode: member names are unique
PASS Element includes ChildNode: member names are unique
PASS Element includes Slottable: member names are unique
FAIL CommandEvent interface: existence and properties of interface object assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface object length assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface object name assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface: attribute invoker assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent interface: attribute command assert_own_property: self does not have own property "CommandEvent" expected property "CommandEvent" missing
FAIL CommandEvent must be primary interface of new CommandEvent("invoke") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CommandEvent"
FAIL Stringification of new CommandEvent("invoke") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CommandEvent"
FAIL CommandEvent interface: new CommandEvent("invoke") must inherit property "invoker" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CommandEvent"
FAIL CommandEvent interface: new CommandEvent("invoke") must inherit property "command" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: CommandEvent"
